Handle container image and layer adding in memory
