+++ pkg/containerutils/containeroptions/addfiletoimagearchiveoptions.go
@@ -4,19 +4,25 @@ import (
 	"github.com/asciich/asciichgolangpublic/pkg/tracederrors"
 )
 
-type AddFileToImageOptions struct {
+type AddFileToImageArchiveOptions struct {
 	// Path to the local file to add:
 	SourceFilePath string
 
 	// Path inside the container image where the source file is added:
 	PathInImage string
 
+	// New name of the new Image after adding the file:
+	NewImageNameAndTag string
+
+	// Overwrite source archive with the updated image archive:
+	OverwriteSourceArchive bool
+
 	// FileMode to set.
 	// Eg. to set it to 0644 you can use: pointerutils.ToInt64Pointer(0644) during initialization.
 	Mode *int64
 }
 
-func (a *AddFileToImageOptions) GetSourceFilePath() (string, error) {
+func (a *AddFileToImageArchiveOptions) GetSourceFilePath() (string, error) {
 	if a.SourceFilePath == "" {
 		return "", tracederrors.TracedError("SourceFilePath not set")
 	}
@@ -24,7 +30,7 @@ func (a *AddFileToImageOptions) GetSourceFilePath() (string, error) {
 	return a.SourceFilePath, nil
 }
 
-func (a *AddFileToImageOptions) GetPathInImage() (string, error) {
+func (a *AddFileToImageArchiveOptions) GetPathInImage() (string, error) {
 	if a.PathInImage == "" {
 		return "", tracederrors.TracedError("PathInImage not set")
 	}
@@ -32,7 +38,15 @@ func (a *AddFileToImageOptions) GetPathInImage() (string, error) {
 	return a.PathInImage, nil
 }
 
-func (a *AddFileToImageOptions) GetMode() (int64, error) {
+func (a *AddFileToImageArchiveOptions) GetNewImageNameAndTag() (string, error) {
+	if a.NewImageNameAndTag == "" {
+		return "", tracederrors.TracedError("NewImageNameAndTag not set")
+	}
+
+	return a.NewImageNameAndTag, nil
+}
+
+func (a *AddFileToImageArchiveOptions) GetMode() (int64, error) {
 	if a.Mode == nil {
 		return 0, tracederrors.TracedErrorf("Mode not set")
 	}
